+++ aws-java-sdk-redshift/src/main/java/com/amazonaws/services/redshift/AmazonRedshiftClient.java
@@ -304,6 +304,7 @@ public class AmazonRedshiftClient extends AmazonWebServiceClient implements Amaz
         exceptionUnmarshallers.add(new ClusterSubnetQuotaExceededExceptionUnmarshaller());
         exceptionUnmarshallers.add(new InvalidClusterParameterGroupStateExceptionUnmarshaller());
         exceptionUnmarshallers.add(new ClusterSubnetGroupNotFoundExceptionUnmarshaller());
+        exceptionUnmarshallers.add(new DependentServiceUnavailableExceptionUnmarshaller());
         exceptionUnmarshallers.add(new SnapshotCopyAlreadyDisabledExceptionUnmarshaller());
         exceptionUnmarshallers.add(new ClusterSecurityGroupNotFoundExceptionUnmarshaller());
         exceptionUnmarshallers.add(new ClusterSnapshotQuotaExceededExceptionUnmarshaller());
@@ -2837,6 +2838,9 @@ public class AmazonRedshiftClient extends AmazonWebServiceClient implements Amaz
      *         Specified offering does not exist.
      * @throws UnsupportedOperationException
      *         The requested operation isn't supported.
+     * @throws DependentServiceUnavailableException
+     *         Your request cannot be completed because a dependent internal service is temporarily unavailable. Wait 30
+     *         to 60 seconds and try again.
      * @sample AmazonRedshift.DescribeReservedNodeOfferings
      * @see <a href="http://docs.aws.amazon.com/goto/WebAPI/redshift-2012-12-01/DescribeReservedNodeOfferings"
      *      target="_top">AWS API Documentation</a>
@@ -2892,6 +2896,9 @@ public class AmazonRedshiftClient extends AmazonWebServiceClient implements Amaz
      * @return Result of the DescribeReservedNodes operation returned by the service.
      * @throws ReservedNodeNotFoundException
      *         The specified reserved compute node not found.
+     * @throws DependentServiceUnavailableException
+     *         Your request cannot be completed because a dependent internal service is temporarily unavailable. Wait 30
+     *         to 60 seconds and try again.
      * @sample AmazonRedshift.DescribeReservedNodes
      * @see <a href="http://docs.aws.amazon.com/goto/WebAPI/redshift-2012-12-01/DescribeReservedNodes" target="_top">AWS
      *      API Documentation</a>
@@ -3449,6 +3456,83 @@ public class AmazonRedshiftClient extends AmazonWebServiceClient implements Amaz
 
     /**
      * <p>
+     * Returns a database user name and temporary password with temporary authorization to log in to an Amazon Redshift
+     * database. The action returns the database user name prefixed with <code>IAM:</code> if <code>AutoCreate</code> is
+     * <code>False</code> or <code>IAMA:</code> if <code>AutoCreate</code> is <code>True</code>. You can optionally
+     * specify one or more database user groups that the user will join at log in. By default, the temporary credentials
+     * expire in 900 seconds. You can optionally specify a duration between 900 seconds (15 minutes) and 3600 seconds
+     * (60 minutes). For more information, see Generating IAM Database User Credentials in the Amazon Redshift Cluster
+     * Management Guide.
+     * </p>
+     * <p>
+     * The IAM user or role that executes GetClusterCredentials must have an IAM policy attached that allows the
+     * <code>redshift:GetClusterCredentials</code> action with access to the <code>dbuser</code> resource on the
+     * cluster. The user name specified for <code>dbuser</code> in the IAM policy and the user name specified for the
+     * <code>DbUser</code> parameter must match.
+     * </p>
+     * <p>
+     * If the <code>DbGroups</code> parameter is specified, the IAM policy must allow the
+     * <code>redshift:JoinGroup</code> action with access to the listed <code>dbgroups</code>.
+     * </p>
+     * <p>
+     * In addition, if the <code>AutoCreate</code> parameter is set to <code>True</code>, then the policy must include
+     * the <code>redshift:CreateClusterUser</code> privilege.
+     * </p>
+     * <p>
+     * If the <code>DbName</code> parameter is specified, the IAM policy must allow access to the resource
+     * <code>dbname</code> for the specified database name.
+     * </p>
+     * 
+     * @param getClusterCredentialsRequest
+     *        The request parameters to get cluster credentials.
+     * @return Result of the GetClusterCredentials operation returned by the service.
+     * @throws ClusterNotFoundException
+     *         The <code>ClusterIdentifier</code> parameter does not refer to an existing cluster.
+     * @throws UnsupportedOperationException
+     *         The requested operation isn't supported.
+     * @sample AmazonRedshift.GetClusterCredentials
+     * @see <a href="http://docs.aws.amazon.com/goto/WebAPI/redshift-2012-12-01/GetClusterCredentials" target="_top">AWS
+     *      API Documentation</a>
+     */
+    @Override
+    public GetClusterCredentialsResult getClusterCredentials(GetClusterCredentialsRequest request) {
+        request = beforeClientExecution(request);
+        return executeGetClusterCredentials(request);
+    }
+
+    @SdkInternalApi
+    final GetClusterCredentialsResult executeGetClusterCredentials(GetClusterCredentialsRequest getClusterCredentialsRequest) {
+
+        ExecutionContext executionContext = createExecutionContext(getClusterCredentialsRequest);
+        AWSRequestMetrics awsRequestMetrics = executionContext.getAwsRequestMetrics();
+        awsRequestMetrics.startEvent(Field.ClientExecuteTime);
+        Request<GetClusterCredentialsRequest> request = null;
+        Response<GetClusterCredentialsResult> response = null;
+
+        try {
+            awsRequestMetrics.startEvent(Field.RequestMarshallTime);
+            try {
+                request = new GetClusterCredentialsRequestMarshaller().marshall(super.beforeMarshalling(getClusterCredentialsRequest));
+                // Binds the request metrics to the current request.
+                request.setAWSRequestMetrics(awsRequestMetrics);
+            } finally {
+                awsRequestMetrics.endEvent(Field.RequestMarshallTime);
+            }
+
+            StaxResponseHandler<GetClusterCredentialsResult> responseHandler = new StaxResponseHandler<GetClusterCredentialsResult>(
+                    new GetClusterCredentialsResultStaxUnmarshaller());
+            response = invoke(request, responseHandler, executionContext);
+
+            return response.getAwsResponse();
+
+        } finally {
+
+            endClientExecution(awsRequestMetrics, request, response);
+        }
+    }
+
+    /**
+     * <p>
      * Modifies the settings for a cluster. For example, you can add another security or parameter group, update the
      * preferred maintenance window, or change the master user password. Resetting a cluster password or modifying the
      * security groups associated with a cluster do not need a reboot. However, modifying a parameter group requires a
@@ -3474,6 +3558,8 @@ public class AmazonRedshiftClient extends AmazonWebServiceClient implements Amaz
      *         your quota, go to <a
      *         href="http://docs.aws.amazon.com/redshift/latest/mgmt/amazon-redshift-limits.html">Limits in Amazon
      *         Redshift</a> in the <i>Amazon Redshift Cluster Management Guide</i>.
+     * @throws NumberOfNodesPerClusterLimitExceededException
+     *         The operation would exceed the number of nodes allowed for a cluster.
      * @throws ClusterSecurityGroupNotFoundException
      *         The cluster security group name does not refer to an existing cluster security group.
      * @throws ClusterParameterGroupNotFoundException