+++ aws-java-sdk-redshift/src/main/java/com/amazonaws/services/redshift/model/transform/AccountWithRestoreAccessStaxUnmarshaller.java
@@ -47,6 +47,11 @@ public class AccountWithRestoreAccessStaxUnmarshaller implements Unmarshaller<Ac
                     accountWithRestoreAccess.setAccountId(StringStaxUnmarshaller.getInstance().unmarshall(context));
                     continue;
                 }
+
+                if (context.testExpression("AccountAlias", targetDepth)) {
+                    accountWithRestoreAccess.setAccountAlias(StringStaxUnmarshaller.getInstance().unmarshall(context));
+                    continue;
+                }
             } else if (xmlEvent.isEndElement()) {
                 if (context.getCurrentDepth() < originalDepth) {
                     return accountWithRestoreAccess;