+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# __1.11.118__ __2017-04-07__
+## __Amazon Redshift__
+  - ### Features
+    - This update adds the GetClusterCredentials API which is used to get temporary login credentials to the cluster. AccountWithRestoreAccess now has a new member AccountAlias, this is the identifier of the AWS support account authorized to restore the specified snapshot. This is added to support the feature where the customer can share their snapshot with the Amazon Redshift Support Account without having to manually specify the AWS Redshift Service account ID on the AWS Console/API.
+
 # __1.11.117__ __2017-04-06__
 ## __Elastic Load Balancing__
   - ### Features